+++ src/EventManager.js
@@ -105,9 +105,11 @@ class EventManager {
             // not using a validator
             handler(event);
           } else {
-            this.validator.validates(event).then(() => {
-              // event passed validation
-              handler(event);
+            this.validator.validates(event).then((success) => {
+              if (success) {
+                // event passed validation
+                handler(event);
+              }
             });
           }
         }