+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.0 - ?
+
+* Add unit tests
+* Bump @superbalist/js-pubsub to ^1.0.0
+* Change `EventValidatorInterface` to return a promise which resolves to a boolean
+
 ## 0.0.1 - 2017-05-03
 
 * Initial release